Several fixes
Org weapons now have their member name.
Dream shield now does render in the pedestal and it's accepted in the slot.
Now you can't interact with pedestals in the SoA if you are in survival.
Optimized Tooltips code a bit.
Set default rotation to the pedestals.
Model rotation and position for the Dream Shield.